+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+1. Added Newton's method root solver.
+1. Updated all functions to accept `f: &impl Fn(f64) -> f64` instead of `f: fn(f64) -> f64`.
+1. Updated the `bisection_method` function to actually track the termination reason.
+1. Updated `once_cell` dependency from `1.19.0` to `1.20.2`.
+1. Updated `numtest` dev dependency from `0.1.6` to `0.2.1`.
+
 ## 0.1.0
 
 1. Initial release with bisection method solvers and basic utilities.